requirements.txt
mysql-connector-python==8.0.28
numpy==1.22.1
opencv-contrib-python==4.5.5.62
Pillow==9.0.0
protobuf==3.19.3
